requirements.txt

streamlit
pandas
requests
Pillow
python-dotenv
google-generativeai